distribution, layer: extend external layer urls to other os
