+++ distribution/pull_v2.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"net/http"
 	"net/url"
 	"os"
 	"runtime"
@@ -133,7 +134,8 @@ type v2LayerDescriptor struct {
 	V2MetadataService *metadata.V2MetadataService
 	tmpFile           *os.File
 	verifier          digest.Verifier
-	src               distribution.Descriptor
+	// TODO(runcom): change this to "d" and incorporate "digest" from above?
+	src distribution.Descriptor
 }
 
 func (ld *v2LayerDescriptor) Key() string {
@@ -290,6 +292,36 @@ func (ld *v2LayerDescriptor) Download(ctx context.Context, progressOutput progre
 		}
 		return err
 	}), size, nil
+}
+
+var _ distribution.Describable = &v2LayerDescriptor{}
+
+func (ld *v2LayerDescriptor) Descriptor() distribution.Descriptor {
+	return ld.src
+}
+
+func (ld *v2LayerDescriptor) open(ctx context.Context) (distribution.ReadSeekCloser, error) {
+	if len(ld.src.URLs) == 0 {
+		blobs := ld.repo.Blobs(ctx)
+		return blobs.Open(ctx, ld.digest)
+	}
+
+	var (
+		err error
+		rsc distribution.ReadSeekCloser
+	)
+
+	// Find the first URL that results in a 200 result code.
+	for _, url := range ld.src.URLs {
+		rsc = transport.NewHTTPReadSeeker(http.DefaultClient, url, nil)
+		_, err = rsc.Seek(0, os.SEEK_SET)
+		if err == nil {
+			break
+		}
+		rsc.Close()
+		rsc = nil
+	}
+	return rsc, err
 }
 
 func (ld *v2LayerDescriptor) Close() {
