+++ distribution/pull_v2_unix.go
@@ -3,17 +3,10 @@
 package distribution
 
 import (
-	"github.com/docker/distribution"
-	"github.com/docker/distribution/context"
 	"github.com/docker/distribution/manifest/schema1"
 	"github.com/docker/docker/image"
 )
 
 func detectBaseLayer(is image.Store, m *schema1.Manifest, rootFS *image.RootFS) error {
 	return nil
-}
-
-func (ld *v2LayerDescriptor) open(ctx context.Context) (distribution.ReadSeekCloser, error) {
-	blobs := ld.repo.Blobs(ctx)
-	return blobs.Open(ctx, ld.digest)
 }
