+++ distribution/push_v2.go
@@ -241,7 +241,7 @@ func (pd *v2PushDescriptor) DiffID() layer.DiffID {
 
 func (pd *v2PushDescriptor) Upload(ctx context.Context, progressOutput progress.Output) (distribution.Descriptor, error) {
 	if fs, ok := pd.layer.(distribution.Describable); ok {
-		if d := fs.Descriptor(); len(d.URLs) > 0 {
+		if d := fs.Descriptor(); fs.Descriptor().MediaType == schema2.MediaTypeForeignLayer {
 			progress.Update(progressOutput, pd.ID(), "Skipped foreign layer")
 			return d, nil
 		}
