--- layer/ro_layer_windows.go
@@ -1,9 +1,0 @@
-package layer
-
-import "github.com/docker/distribution"
-
-var _ distribution.Describable = &roLayer{}
-
-func (rl *roLayer) Descriptor() distribution.Descriptor {
-	return rl.descriptor
-}
